mcp-stack/server/app/main.py
"""Main FastAPI application for the MCP Server."""
import os
import logging
from pathlib import Path
from typing import Any, Dict, Optional

from fastapi import FastAPI, HTTPException, Request, status
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
import strawberry
from strawberry.fastapi import GraphQLRouter

from .config import settings
from .graphql.schema import schema
from .graphql.data_schema import DataQuery as DataQueryType
from .services.data_service import DataService

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Initialize FastAPI app
app = FastAPI(
    title="MCP Server",
    description="Model Context Protocol Server with GraphQL API",
    version="0.1.0",
    docs_url="/docs",
    redoc_url=None,
)

# Create data directory if it doesn't exist
DATA_DIR = Path(settings.DATA_DIR)
DATA_DIR.mkdir(parents=True, exist_ok=True)

# Initialize services
data_service = DataService(data_dir=DATA_DIR)

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=settings.cors_origins_list,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Create GraphQL router
graphql_app = GraphQLRouter(schema)
app.include_router(graphql_app, prefix="/graphql")

# Health check endpoint
@app.get("/health")
async def health_check() -> Dict[str, str]:
    """Health check endpoint."""
    return {"status": "healthy", "version": app.version}

# Add data query router
data_query_router = GraphQLRouter(
    schema=strawberry.Schema(query=DataQueryType),
    context_getter=lambda: {"data_service": data_service}
)
app.include_router(data_query_router, prefix="/data")

# Error handling
@app.exception_handler(HTTPException)
async def http_exception_handler(request, exc):
    """Handle HTTP exceptions and return consistent error responses."""
    return JSONResponse(
        status_code=exc.status_code,
        content={"detail": exc.detail},
    )

# Startup event
@app.on_event("startup")
async def startup_event():
    """Initialize application services on startup."""
    # Ensure data directory exists
    data_dir = Path(settings.DATA_DIR)
    data_dir.mkdir(parents=True, exist_ok=True)

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(
        "app.main:app",
        host=settings.HOST,
        port=settings.PORT,
        reload=settings.DEBUG,
    )
