mcp-stack/server/requirements.txt
# Core dependencies
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
python-multipart==0.0.6

# GraphQL
strawberry-graphql[fastapi]==0.166.0

# Data processing
numpy==1.24.3
pandas==1.5.3
pyarrow==10.0.1

# Logging
loguru==0.7.0

# Pydantic (must be <2.0.0 for compatibility with FastAPI 0.95.2)
pydantic==1.10.13

# Required by FastAPI
anyio==3.7.1
click==8.1.3
h11==0.14.0
httptools==0.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.5.0
watchfiles==0.19.0
websockets==11.0.3

# Text processing
textblob==0.17.1
rake-nltk==1.0.6
nltk==3.8.1
python-json-logger==4.0.0
